[#861] Fix $ZATRANSFORM replies for 1st parm numeric expressions that have no string component
The op_zatransform.c module did not ensure the existence of a string component om the input value
when one is required. When computed expressions were used as the input value (e.g. 1-1 or 1&0), the
input value had no string component so was referencing garbage. The fact that it worked at all in
r1.30 or the relevant GT.M releases is surprising. The issue was eventually found by fuzzing tests.