[#604] [V63013] Address clang-tidy pipeline failure
This commit addresses the clang-tidy pipeline failure by updating the clang-tidy warning files
and by making code changes to address the following clang-tidy warnings:

`dbfilop.c:warning: The right operand of '==' is a garbage value [clang-analyzer-core.UndefinedBinaryOperatorResult]`

The above warning is fixed by initializing the `save_errno` variable to 0 in `sr_unix/dbfilop`.

`t_end_sysops.c:warning: Value stored to 'chain1' is never read [clang-analyzer-deadcode.DeadStores]`

The above warning is fixed by putting all usages of the `chain1` variable in `sr_port/t_end_sysops.c`
behind `DEBUG_ONLY()` macros. This variable was only read by an assert which is why it generated the
clang-tidy warning on pro builds.

`gvcst_init_sysops.c:warning: Value stored to 'save_errno' is never read [clang-analyzer-deadcode.DeadStores]`

The above warning is fixed by removing a redundant line in `sr_unix/gvcst_init_sysops.c` that set
the variable `save_errno`.

The other clang-tidy warnings in the pro failure were 5 now fixed warnings, changes to the number
of padding bytes in 2 existing warnings and a new false alarm warning in `sr_port/gvt_inline.h`
about the following line:

`keyend = gvkey->end;`

This supposedly dereferences a null pointer but `gvkey` is a parameter that will never be NULL.

After fixing the pro failure, it also failed with a dbg build. This involved 4 fixed warnings
that are now removed from the debug warning file, changes to the padding bytes in 2 existing
warnings and a new warning in `sr_port/repl_comm.c`. The new warning in `repl_comm.c` became
a `value stored is never read` when the field was initialized to 0 so that change was reverted
and the warning added to the warning file. It only shows up in dbg.