[#604] [V63013] Fix incorrect merge conflict resolution in sr_port/op_fnquery.c
This commit fixes an incorrect merge conflict resolution in `sr_port/op_fnquery.c`. Specifically,
the order of conditions in an `else if` was changed in V6.3-013. In upstream, there are 2
conditions: `(nexti == sbscnt)` and `(0 == arg1->str.len)`. In YottaDB, there is a 3rd condition:
`!is_sqlnull`. In upstream, the order of the 2 conditions was reversed in V6.3-013 and this
conflict was originally resolved by reversing the 2 conditions, leaving `!is_sqlnull` as the 2nd
of the 3 conditions. This caused test failures in the subtests `simplethreadapi/gvnlvnnodenext`,
`simplethreadapi/nodenext` and `mvts/mvts` with stack smashing and a signal handling core. These
failures showed up 100% of the time. With this change reversed, those tests pass 100% of the time.

There was also a second incorrect merge conflict resolution where the change to the variable
`history` from 18e79576249d4c3cfe8d6356447c6b26abcdcccc was accidentally undone. On ASAN builds,
this caused test failures in the same subtests as the other issue plus the simpleapi versions
of the 2 simplethreadapi tests. These failures are fixed once the bounds of `history` are restored
to `MAX_LVSUBSCRIPTS + 1` not `MAX_LVSUBSCRIPTS`.