Fix r136/ydb864 test failure by adding a white box test case to ensure a static process ID in the tempfile
For a `MUPIP BACKUP -noonline`, the temp file name contains the process ID.
Since the number of characters in the process ID can differ, this can cause
a backup in the r134/ydb864 test involving a 234 character directory name
to unexpectedly succeed when the pid is 9999 or less which then causes the
test to fail. The fix for this is to add a white box test which sets the
process id part of the temp file name to 99999. This ensures that the
temp file name will always be a predictable length.