[#879] Stop looping in source server on invalid secondary system name after applying GTM-9157
Prior to GTM-9157, when the source server was told to connect to a system name that could not be resolved or if
it became unresolvable over time, the source server was terminated. Applying GTM-9157 (as part of our merge of
V63012 into YottaDB), caused a few different issues:

1) The hardspin loop waiting for the connection had no check for outofband type issues except at the end.
2) The hardspin loop also had no wait in it. Only the path with link resolution had a sleep in it.
3) Ditto 1 & 2 for the softspin loop.

Because of this, if a bad dns resolution happend, unless the shutdown signal came in before the hardspin loop
finished (in just a few milliseconds), we would go into an eternal retry loop that did not obey the source server
shutdown signal. Similarly, because neither the hardspin or softspin loops were sleeping, the server ran at 100%
trying to connect over and over many times per second. A side effect of this (since the loops were assuming they
were sleeping, the "trying for xxxxxxx seconds" type messages were way out of line saying that it had been
waiting for over 10K seconds when it had only been running around a minute or two.

The fix is to make sure that if the resolution of or connection to a given secondary fails, regardless of which
part failed, is to make sure that both paths do two things: (1) do the appropriate sleep between attemtps and
(2) after the sleep, check for pending shutdowns or other interrupts with a call to gtmsource_poll_action(FALSE).
Previously, resolution issues caused an immediate error that terminated the source server.

Note, the test of this functionality is done by the new v63012/gtm9157 test.