[#604] [V63013] Remove assert from cmd.c that caused ZI to crash dbg builds
The `manually_start/largelibtest` test started failing on dbg builds after integrating
V6.3-013 into YottaDB. Upon investigation, we discovered that the compiler changes
in V6.3-013 caused `ZI` commands in some VistA M routines that are part of the test,
which are supposed to be calling a `ZINSERT` command that exists on some other M
implementations, to call `ZINVCMD` which instantly crashes a dbg build with an assert
failure and core. After examining it on various versions, the commands `ZI`, `ZIN`,
`ZINV`, `ZINVC`, `ZINVCM` and `ZINVCMD` all cause assert failures on dbg builds for
upstream versions V6.3-013 through V7.0-002. For versions prior to V6.3-013, only
the command `ZINVCMD` caused an assert failure. The specific assert that failed
was for `m_zinvcmd != cmd_data[x].fcn` which checked that the command entered was
not `ZINVCMD`. Since VistA code written for other M implementations should not
cause a crash on YottaDB and since removing the assert restores the pre-V6.3-013
behavior (a `%YDB-E-INVCMD` error message), this commit removes the assert.