[#604] [V63013] Make equivalent change to YDB repo version of sr_unix/ydb_tls_interface.h
This commit makes a change to `sr_unix/ydb_tls_interface.h` to rename `GTMTMLS_OP_NOSHUTDOWN`
to `GTMTLS_OP_NOSHUTDOWN`. This change was made as part of the review of YottaDB/Util/YDBEncrypt!12.

This commit also reverts the copyright changes in `sr_unix/ydbcrypt_interface.h` since there are no
material changes to the file in V6.3-013.