[#886] Setting $ZSTRPLLIM  always does a minimum size check and setting $ZSTRPLLIM below the minimum sets it to the minimum
There are two issues here:
1. If $ZSTRPLLIM is being set for the first time or is being increased, then the minimum size
check is bypassed. This allows values less than the minimum to be set which should not happen.
2. The release note for GTM-9238 for V63012 [see it at YDBTest#420] states that if the value being
assigned to $ZSTRPLLIM is less than the minimum number (defined to be 100K) then the number
specified is increase by the minimum.

When considering a fix for this, we decided that this functionality that adds the minimum to the
setting can cause some odd results - set the limit to 100K and that works but reduce it to 50K
and get 150K instead. So we've made a further change to set values that are below the minimum
to the minimum to avoid value changes contrary to what the user was trying to accomplish.