[#893] Raise $ZSTRPLLIM minimum value from 100K to 150K
The 100K lower limit (which is what value is set to for any value less than 150K) is increased
to 150K to allow minimal test programs to run. Previously the 100K minimum, when used, would
cause even a simple ZWRITE command of a single value to get an STPCRIT error under most
circumstances.