[#888] Add $ZGLD as a synonym to $ZGBLDIR
Since GT.M V6.3-013 got merged, $ZGLD ISV stopped being accepted as a
valid abbreviation for $ZGBLDIR ISV. The reason is that this release of
GT.M strongly tightened what's a valid abbreviation or not based on the
documentation. However, it still accepts short forms (e.g. $ZG) as long
as there is no collision with other names.

We know of users who use $ZGLD in production code. This commit is to
accomodate them. There may be future commits like this to add back
abbreviations commonly used by end users that are not officially in the
GT.M/YottaDB manuals.