[#891] Make SUSE a Supported distribution
In `getosid()` treat openSUSE Leap as equivalent to SUSE Linux Enterprise Server (SLED) and Desktop (SLED).

This was bench-tested tested with the `--dry-run` option, as there is no YDB distribution tarball for SUSE Linux yet. Since the distribution tarball requires this change, there is a circular dependency which requires this to be merged first.