[YottaDB/Tools/YDBCMake#5] Fix errors in previous commit 988c0c40f394d92069fc2e15e9be6fe444f049d8
It turns out I didn't copy the Posix block correctly; and missed the
then clause of the if/then/else. This is now fixed.

As a fortunate side effect, I found that the error message for failure
was incorrect in the directory description, and that is now fixed.

Tested manually on a clean Ubuntu 22.04 docker container, and YDBTest
sudo test suite ran.