[#926][Pipeline] Build dockers on master only; cleanup
Due to https://about.gitlab.com/blog/2022/03/24/efficient-free-tier/, we
need to reduce our unneeded storage. The big issue in the YDB repo is
with the docker images, which get created all the time. This is not
needed, and they will only get created after a merge to master now.

Other notable changes:

- Stages changed from build, push to build, docker-create, docker-push
- docker-create artifacts is only for 1 day, as they get uploaded to
  Docker Hub right away.