[#927] Cleanup left over ftok semaphore after DBFILERDONLY error
The initial YDB#927 fix (!1215 in commit https://gitlab.com/YottaDB/DB/YDB/-/commit/be80269da89ace3b71867d563f45385e30a407ff)
had a problem with an intermittent failure in the test created for GTM-9278 (YDBTest#450). The issue was an assert failure in
ftok_sem_get_common.c and the assert was that the value of the global variable ftok_sem_reg (the region the ftok semaphore is
for) is NULL. In normal operation, this variable is cleared when initialization is complete. But if some really weird errors
occur in delicate places, this value could be left non-zero and then fail this assert when we go to open another dbfile.

This support does two things:
1. Move the code added by !1215 to an earlier block so we only check if deleting the semaphore is proper if we have grabbed
the semaphore (thus holding the lock).
2. Added a cleanup for ftok_sem_reg after we've cleared the flags pointed to by udi which is a safe place to do this.