[#927] Cleanup left over ftok semaphore after DBFILERDONLY error
The v63013/gtm9278 test ran into a problem where the ftok semaphore was left over after a DBFILERDONLY
error message. The test actually resets the global directory so it points at an empty directory instead
of a database file. This causes some errors that put out additional information now than they did
previously. But it also usually left the ftok semaphore laying around which cause a test failure.
Sometimes, it did not error out but we were unable to determine why so just fixed the issue causing
the error to avoid the issue entirely.

This fix adds a new flag in the udi (unix_db_info structure) to indicate that this process created the
ftok semaphore. If so, the semaphore is removed on an error exit during database initialization.