[Pipeline] Change commit-verify due to local runners
We are deploying local runners for the pipelines. We found out that by
default Gitlab mounts a volume with local repositories that may be
already initialized; and it's possible that "upstream_repo" is already
configured. This commit changes the script so that upstream_repo is
re-used if it already exists.

An alternative is to always clean the docker cache after each set-up
(https://docs.gitlab.com/runner/executors/docker.html#clearing-docker-cache)
but this is inefficient as it will lengthen execution times.