[#765] [#924] ydbinstall returns error if UTF-8 mode install failed; also terminates after processing command line if not run as root, unless --dry-run is specified
`ydbinstall.sh` previously reported success for installing YottaDB in UTF-8 mode, even though it actually failed. Now, the installer will correctly report its success. [#765]

The check in `ydbinstall` as to whether it is running as root is made after processing the command line. Unless `--dry-run` is specified, it terminates with an error at that point, rather than after further processing and downloading. [#924]

Fixes to issues found in testing.