[#945] ydb_env_set returns non-zero status on MUPIP CREATE errors
When MUPIP CREATE reports errors, sourcing `ydb_env_set` returns a non-zero status to the shell, and reports the location of the MUPIP CREATE error output. Previously, it silently ignored MUPIP CREATE errors and reported a zero status.

Force NOASYNCIO in `gdedefaults` as the default database block size of 2048 bytes is incompatible with ASYNCIO in filesystems with 4096 byte blocks.

Ensure that if $ydb_gbldir / $gtmgbldir exists, it is indeed a global directory file and that database files pointed to by segments in the global directory are indeed database files.

Responses to two rounds of review comments