[#946] Change repl_inst_dump.c to print jnlpool size as a 64 bit value instead of just 32 bit
In V6.3-014, the jnlpool_size field in the replication instance file (jnlpool_ctl_struct) was
changed from a 32 bit value to a 64 bit value but the code in repl_inst_dump.c was not changed
to support the new larger size. This fix changes how the field is printed so the 64 bit value
is shown fully instead of truncated if the journal pool size exceeds 4GB.

Note - this functionality is tested by v63014/gtm7628 (which depends on it).