[#904] $ZY should not be considered a valid abbreviation
GTM-9259, part of GT.M V6.3-013, integrated into YottaDB for r1.36, has
the following change:

> The GT.M compiler checks Z* keywords except for deviceparameters for
> the length provided by the source code and accepts a leading subset,
> with a small set of legacy abbreviations. This means that if you use a
> short form that's not unique, GT.M evaluates it as the alphabetically
> first keyword or legacy abbreviation. FIS recommends using at least four
> characters in stored code, but ZBIT* functions and deviceparameters with
> a leading "NO" may need more. For standard keywords, the compiler
> requires correct spelling for the full keyword through up to eight
> characters and recognizes standard abbreviations. Previously the
> compiler ignored characters following matches in its internal tables
> which caused surprise if there were similar keywords or typos - in
> particular it may reject misspellings it accepted in prior releases.
> (GTM-9259).

As part of discussion around #925, we decided to keep that change
integrated into YottaDB, since this is how GT.M historically has worked,
however, for $ZY ISVs and functions, we didn't want $ZY itself to be a
considered a correct abbreviation (it resolves to `$ZYERROR`), since all
of $ZY* is supposed to be YottaDB commands, which need to be fully
specified. As part of making that change, we found out that our
documentation said that $ZYER was a valid abbreviation for $ZYERROR and
$ZYRE for $ZYRELEASE. Furthermore, the test system has code that relies
on $ZYERR.

This commit disables wildcard matching for any names starting with $ZY
(functions, ISVs, or commands); and adds $ZYER, $ZYERR and $ZYRE as
valid abbreviations.