[#931] Build AARCH64/ARM64 images on pipeline + clang-tidy
Due to https://github.com/docker/buildx/issues/1331, we could not use
docker buildx to build AARCH64 images. Instead, what we do is create the
two of them (AMD64 and AARCH64) seperately on an AMD64 and an AARCH64
machine, tag and push them separately, and then download them back and
combine them using `docker manifest`.

The main other item done here is run clang-tidy on our AARCH64 machine,
but only on master, as it takes too long to run (about 35 minutes). We
needed seprate reference files for clang-tidy warnings, as our source
code is differnt for AARCH64 (we remove the x86_64 folder and add the
aarch64 folder).

As a side effect of needing to build the docker images a few changes
were needed:

1. The `Dockerfile-debian` performed tests inside of the docker file,
   and one of these tests in YDBgo failed on AARCH64 due to the fact
   that it is a speed test, and the system we use is kind of slow. Tests
   are not appropriate in the building process in any case, so I decided
   to remove them; we have pipelines in each of the other repos to run
   tests. The exception is the test for the Perl wrapper, as we don't
   test that anywhere else.
2. The upsteam image `rockylinux/rockylinux` did not have AARCH64 docker
   images. This was identfied in earlier work here:
   https://github.com/WorldVistA/docker-vista/commit/792b3b9e1dc2b276fd70f1f2e049db5413ade67c.
   The solution is similar: change to `rockylinux:8.7`.
3. We do not support RHEL-like systems (which Rocky Linux is an emulation of)
   on AARCH64. `ydbinstall.sh` rejected installing YottaDB on Rocky Linux.
   However, rather than make an exception in the building process for
   Rocky Linux, I utilized the ability of `ydbinstall.sh` to recognize
   the fact that we just compiled YottaDB on that platform, and to
   install the compiled version. That entailed a small change to the
   Dockerfile for Rocky Linux to add a CMake configure time file to it,
   which is not present at the second stage.

This is the testing that was done:

1. All development test images were pushed to hub.docker.com/r/samindcon
2. Test that `:latest-master` (each commit) and `:latest` (each tag) images get created
3. All `:latest-master` images were tested on both AMD64 and AARCH64.
   `:latest` Ubuntu image tested on both AMD64 and AARCH64.