[#779] [V63014] Fix incorrect merge conflict resolution in sr_unix/gt_timers.c when SIGSAFE external call steals timer handler
In b948dede, as part of merge conflict resolution of the GT.M V6.3-014 changes
into YottaDB, two lines of code that previously were under a if (sig_handler_changed)
check were inadvertently moved outside that check. The result of this was that
even if SIGSAFE was specified (which is a guarantee by the user that the external
call does not change the signal handlers), we checked to see if any timer pops
had been lost. This commit moves those two lines back inside the conditional so
SIGSAFE external calls avoid that overhead.