[#969] When --utf8 is not specified ydbinstall does not check for libicu.io
An incorrect check as to whether `$ydb_utf8` was the empty string erroneously caused `ydbinstall` / `ydbinstall.sh` to report an error and exit without installing YottaDB even when `--utf8` was not specified on the command line and `libicuio.so` was not found. The correction is to check for `libicu.io` only when `$ydb_utf8` is `"Y"`.