[#931] Build AARCH64/ARM64 images on pipeline
The previous commit (93a949d21606655f007b9231e35a5dfe8c8513e8) forgot to
create a manifest file for the `-base` images.

This effectively caused the `base:latest-master` to not be built; and
resulted in the downstream pipelines to fail when when tried to pull
`base:latest-master` when on AARCH64, as the original image was AMD64.