[Pipeline] Remove tags from clang-tidy-amd64 job
This fixes an issue in the previous
93a949d21606655f007b9231e35a5dfe8c8513e8 commit.

This commit used tags; but most of our runner instances in the gitlab
fork accounts are not tagged. Only the upstream pipeline runner
instances are tagged.

Therefore for all jobs running on master only, it's okay to tag them.
For pipelines not running on master, they should not be tagged, as if
the job is tagged it won't find a pipeline to run.

The original motivation behind the issue was to restrict jobs to run
only on the main AMD64 runner; but I found a much simpler way: Using
https://docs.gitlab.com/ee/ci/runners/configure_runners.html#set-a-runner-to-run-untagged-jobs,
set our AARCH64 runner to only accept jobs destined for it, not as a
general runner. This setting is not okay for our AMD64 runner, but is
what we desire for our AARCH64 runner.