[YottaDB/Util/YDBDevOps#9] Update Dockerfile to work on Ubuntu 22.04 AARCH64
Also, since we removed support for Ubuntu being considered supported on
AARCH64, we add a piece of information (the `build_os_release` file)
so that ydbinstall will realize that we compiled this from source and
it's okay to proceed. See commit
93a949d21606655f007b9231e35a5dfe8c8513e8 for more details.

Dockerfile manually tested on our AARCH64 machine to ensure it works.