[#978][#985] Octo/Rocto + missing netbase
[#978] Octo has been added to the Debian Dockerfile. Rocto is started on
port 1337 with username/password ydb/ydbrocks. The set-up on Rocto is
similar to https://gitlab.com/YottaDB/DBMS/YDBOcto/-/blob/master/Dockerfile.

A sample invocation after this is merged is `docker run --rm -it -p 1337:1337 download.yottadb.com/yottadb/yottadb-debian:latest-master`.

Of note, since we needed to add multiple services to the
`Dockerfile-debian`; the entrypoint for the docker file was changed
from a direct call to a shell to a bash script which can do
some stuff and then transfer control to the main process via `exec`.

[#985] GT.CM requires `/etc/protocols` which is provided by `netbase`
package.

There are some whitespace changes to ensure that all spaces in
Dockerfiles don't contain tabs.