[#979] MUPIP FTOK properly reports FTOK information for repl instance files not in current directory as well as deals with large (> 255 byte) database filenames
Currently, when filenames are processed from the command line in a MUPIP FTOK command, when
a repl instance file is specified along with a -journal pool or -recvpool option, we open
the instance file but then for the actual report, we use parse_file() to expand the repl
file name then throw away its path before running FTOK() on it. This then means that the
ftok and file-id values displayed are always for the file in the current directory (if it
exists). The YottaDB extension to this command allows multiple replication instance files
to be specified. Because of the stripping of the path prior to doing the FTOK on the file,
all of the files will have the same FTOK value and file-id.

This change removes the parse_file call and lets the FTOK command operate on the same
file we just opened so the correct path/name/ftok/fileid information is presented.

This change also removes the usage of strncpy() in this module. It was improperly used
and caused SIG-11 when a greater than 255 byte path/name was passed in from the command
line.