[#986] Enhance $$^%PEEKBYNAME()/$ZPEEK() to allow access to unix_db_info/UDI
For $$^PEEKBYNAME() utility: Adds the capability to use unix_db_info.* structure fields. In general, these are fields
that contain information about the IPCs used and other database identifying attributes.

For $ZPEEK() function: The new UDIREG keyword allows access to the UDI structure to obtain additional information
about the database - specifically the unix_db_info struct. The first part of this struct is the sgmnt_addrs structure
(for which access exists via CSAREG keyword) but the UDIREG keyword gives access to the fields after sgmnt_addrs which
describe the IPCs and filesystem information.

This option was added to satisfy a need in testing to retrieve the ftok_semid field from this
structure for example. This addition is used/tested in v63014/gtm8800.