Added check for copyright-only changes to `commit-verify.sh`
Any YottaDB file that is modified must have its copyright notice updated to reflect the year of the latest change. If, however, the code changes in such a case are reverted, it is possible that the copyright notice is omitted from reversions. This results in an orphaned copyright notice revision, such that the only modification to such files is the copyright notice itself.

Such copyright-only changes are undesirable and easy to overlook. So, this commit adds an automatic check for such changes to `commit-verify.sh`. Now, this script checks for files modified by new commits that contain only copyright notice changes and issues an error and fails the script when it finds any.