[DEBUG-ONLY] [YDBTest#485] Modify an assert in mlk_lock() to allow better testing of GTM-9332
When the pctl->ctl->wakeups counter overflows, it rolls over to 0. But if mlk_lock() returns
this value (which it does), this is the success indicator for mlk_lock() even if the lock
was NOT obtained. But in a debug version, there is an assert for a non-zero value that prevents
us from actually seeing the lock-not-obtained issue. This change adds an extra clause to the
assert so that it also passes if the WBTEST_LCKWAKEOVRFLO white-box test is active. This allows
our v63014/gtm9332 test to fail properly showing the situation when running with a version
based on V63014 (i.e. YottaDB r1.36) that has GTM-9332 reverted.