[YottaDB/DBMS/YDBOcto#946] [#990] YDBAIM region has BG access method and journaling ON; OCTO DDL has less redundant data
Create YDBAIM region with access method, journaling & replication to match the DEFAULT region. Previously, it was an AutoDB region created on demand, with the MM access method, no journaling and no replication. As noted in https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/946#note_1285591279 deleting and recreating the YDBAIM region during crash recovery could result in inconsistencies, and this was the preferred alternative.

Move creation of YDBJNLF region from %YDBJNLF to %YDBENV. Since it was agreed to not journal the region (https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1320#note_1293745665), it remains an AutoDB region, i.e., sourcing `ydb_env_set` creates the region in the global directory; it's just the database file that is created on demand.

The ^%ydbJNLFRECTBL global / YDBJNLFRECTYPE table have the journal format (44) as the first subscript rather than the label. So there will only be one subtree / table for each journal format. This is backwards incompatible change, but is considered acceptable since %YDBJNLF is field test software.

The YDBJNLFRECTYPE table column describing the table type is renamed to `tbltype` from `tbl` and the DDL has a comment added.

Some minor formatting cleanup is included.