Fix issue with checking copyright-only changes in ci/commit_verify.sh script
In the YDBGUI project, a file which contains deletions only, and where
the copyright was correctly updated, was erroneously claimed to be a
copyright only change. This is a regression in the change in
d418add2fbd699261287b0e2ec09896c08b7b41b. That commit only looked for
additions as
signifying a non-copyright only change, but we can also have deletions.

This is fixed by using `git diff --numstat` to count the additions and
deletions, and comparing this against the Copyright line changes, which
must be counted manually.