[#997] Fix ydbinstall header check logic for gpgme.h on ARMV6L (causes build failure)
See https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1337#note_1441727552 for background.

On ARM6, using the C header checker for "gpgme.h" resulted in the
following error, which does not happen on any other platform:

```
In file included from foo.c:5:
/usr/include/gpgme.h:111:2: error: #error GPGME was compiled with _FILE_OFFSET_BITS = 64, please see the section "Largefile support (LFS)" in the GPGME manual.
  111 | #error GPGME was compiled with _FILE_OFFSET_BITS = 64, please see the section "Largefile support (LFS)" in the GPGME manual.
      |  ^~~~~
```

Turns out in YottaDB and YDBEncrypt, we always set `_FILE_OFFSET_BITS`
to be 64; the simple solution is used here, and `_FILE_OFFSET_BITS` is
set to 64 if not previously defined.