Fix commit_verify.sh to handle mode change
`commit_verify.sh` said that files needed copyright change if the mode
was changed (and not the contents); and also complain that ONLY a
copyright change was made if you actually followed its instructions. It
should not require copyright notice for mode-only changes (like removing
the execute bit).

This is done but checking that both additions and deletions are zero.

Also, another line was modified to ensure it runs in Bash Arithmetic
context `(( ))`. `( )` is a string context.