[Pipeline] Split the plugin install step so we can easily see failing steps
In various instances, it was hard to see which step failed due to the
fact that all steps were grouped together, to reduce the size of the
image due to the use of the overlay file system. In this case, this is
not really necessary as the files that get installed don't overwrite
each other, so the sum of the installs will be close to the final file
size. I verified this, and the final image is still 2.79GB, which was
the previous size.

Steps in https://yottadb.com/product/get-started/ manually tested
successfully.

In addition:
- The pip install step for the testing artifacts for the Python plugin
  was inadvertently left in from a previous commit that tried to clean
  this up (93a949d21606655f007b9231e35a5dfe8c8513e8). It has been
  removed as it's not needed.
- ydb_dist is manually set rather than having to call the heavy-weight
  `ydb_env_set` for each install step. Only the Perl install has
  `ydb_env_set` as it needs a database for testing.