[#1025] %YDBJNLF runs in a Docker container without needed /proc filesystem permissions
A non-root process in a container may not have write permission to write to the `/proc` filesystem. That $JOB of a process in a container may not correspond to its host pid, `/proc/self/fd/2` adds a further obstacle to explicitly opening stderr, since YottaDB does not provide a standard M device for stderr.

The solution is to check whether the process can OPEN `"/proc/"_$JOB_"/fd/2"`, and if it cannot, to use the saved value of $IO when the routine was invoked for a ZSHOW output, and to log $ZSTATUS in the syslog using $ZSYSLOG(). Furthermore, if the error occurs within trigger code, the ZSHOW output goes to $PRINCIPAL, since the local variable with the saved value of $IO is not accessible.