[#88] Fix post-merge pipeline failure due to omission of libreadline-dev[el] packages
This pipeline (https://gitlab.com/YottaDB/DB/YDB/-/pipelines/1063833320)
failed due to the absence of libreadline header files, which are now
required to compile YottaDB; readline itself is optional at runtime, as
we can do all user operations without it.

This commit adds the dev packages to the compile stage and the runtime
packages for the release stage.

In additions, the README has now been updated to add libreadline-dev[el]
packages as requirements to compile YottaDB.