[#88] Assert in `recall 0` in readline mode
Reported here
https://gitlab.com/YottaDB/DB/YDBTest/-/merge_requests/1758#note_1668130811:

```
8993bea4b53a:~> setenv ydb_readline true
8993bea4b53a:~> $ydb_dist/yottadb -direct

YDB>recall 0
%YDB-F-ASSERT, Assert failed in /Distrib/YottaDB/V999_R999/sr_unix/readline.c line 546 for expression (FALSE)
```

The recall logic fell to an assert designed to catch all other cases
which are not handled. This happened in this case because we did not
consider `0` a number to be tested.

This is now fixed by restructuring the if/else logic after a non-zero
return from a readline_is_recall() call to mirror the if/else logic
inside readline_is_recall(). An automatic side effect of this is that 0
is considered a valid numeric case.