[DEBUG-ONLY] [GTM-9324] Remove errant assert that trips when using ZSTEP while ZBREAK is restricted
Though the release note for GTM-9324 says that it made ZSTEP restricted whenever ZBREAK was restricted,
this is not actually true. There is an assert added to op_zstep.c that indicates that it should not be
run when ZBREAK is restricted. This assert trips when ZSTEP is used while ZBREAK is restricted in a
debug build. There is no raising of the RESTRICTEDOP error in ZSTEP. If ever that is desired, it needs
to be added. But the assert that verifies ZBREAK is not restricted is bogus and wrong so it is removed.

This change is tested by sudo/gtm9324 in the second half of the test where we verify when ZBREAK is
restricted, we can still use ZSTEP with no assert fail. Note, we don't restrict against running in pro
as it still passes.

Also in this commit, we remove an errant assert in op_zstepret.c that indicates it should not be run
when the BREAK command is restricted but while the BREAK command is the default value for $ZSTEP and
is run whenever the next appropriate line is encountered, it can be other commands, such as in the
test case added in sudo/gtm9324 for this condition and to verify it does not assert fail.