[#1053] Compiling YottaDB if ydb_chset="UTF-8" fails
Problem
-------
It fails with the following error (courtesy of `make VERBOSE=1`):

```
/usr/bin/cmake -D ydb_dist=/home/sam/work/builds/YDB -D ydb_routines=. -D ydb_gbldir=/home/sam/work/builds/YDB/gtmhelp -D mumps=/home/sam/work/builds/YDB/yottadb -D "args=-run;GDE" -D input_file=/home/sam/work/builds/YDB/gtmhelp.in1 -P /home/sam/work/gitlab/YDB/sr_unix/mumps.cmake
%YDB-E-INVOBJFILE, Cannot ZLINK object file /home/sam/work/builds/YDB/GDE.o due to unexpected format
%YDB-I-TEXT, Object compiled with CHSET=M which is different from $ZCHSET
CMake Error at /home/sam/work/gitlab/YDB/sr_unix/mumps.cmake:43 (message):
  Command </home/sam/work/builds/YDB/yottadb -run;GDE
  INPUT_FILE;/home/sam/work/builds/YDB/gtmhelp.in1 > failed with result
  ='253'
```

The reason it fails is that CMakeLists.txt runs `mumps` commands to compile various objects, and these are always compiled in "M" mode and "UTF-8" mode. However, when we run the command (like above), we do not set `ydb_chset`, and as a result, we use whatever `ydb_chset` is currently in the environment.

Solution
--------
1. Specify that we are in M mode vs UTF-8 mode for each compilation. This makes
   no difference, but is better just to be explicit about it. No more "" mode.
2. For each part where we run code, specify the ydb_chset for the code running.