[#1046] Process with ydb_readline=1 terminated by $ZTIMEOUT requires stty sane to regain sanity
From the issue description:

After `ydb_env_set`, without `ydb_readline` the terminal of a process terminated by $ZTIMEOUT is in a sane state:

```
$ unset ydb_readline
$ yottadb -dir

YDB>set $ztimeout=.5

YDB>Error occurred: 150384408,+1^GTM$DMOD,%YDB-W-ZTIMEOUT, Time expired
$ # Terminal is in a sane state
```

With `ydb_readline` the terminal requires stty sane before it is usable:

```
$ export ydb_readline=1
$ yottadb -dir

YDB>set $ztimeout=.5

YDB>Error occurred: 150384408,+1^GTM$DMOD,%YDB-W-ZTIMEOUT, Time expired
$ # Terminal required stty sane before I could use it.
```

The issue does not mention two items, necessary for this bug to
manifest:

1. This happens only on Bash. CShell apparently restores echo when
   control is returned to it.
2. The error trap has to contain a "halt" so that we exit during the
   processing of the error trap.

This is fixed by unsetting terminal settings for Readline after a
SIGALARM is received (outofband == ztimeout).

_Note that in case of $ZTIMEOUT where it does NOT kick us out to the
shell, we will come here too but it does not hurt to do this cleanup
because $ztimeout will cause the current input line to stop being read
and a new input to be read after the M vector in $ztimeout is invoked.
The act of reading new input will cause readline re-initialization and
so there is no issue. The fact that $ztimeout in the middle of an input
line causes current the first half of the input line to be discarded is
a pre-existing issue even with ydb_readline=0 and so that pre-existing
misfeature is not being addressed as part of !1448. See
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1448#note_1706029937
for an example._